lyrics command
Bot can grab lyrics using the GENIUS API
