+++ package.json
@@ -26,6 +26,7 @@
     "dotenv": "^10.0.0",
     "ffmpeg-static": "^5.1.0",
     "fs-extra": "^10.1.0",
+    "genius-lyrics-api": "^3.2.0",
     "node-fetch": "^2.6.1",
     "node-html-parser": "^5.4.2-0",
     "node-ts": "^5.1.2",
